Fixbug: Calculate percentagem to be float

+++ app/classes/v1/dashboard/echart/faults.rb
@@ -2,7 +2,7 @@ class V1::Dashboard::Echart::Faults
   
   def initialize(count)
     config = ::V1::UserConfig.new.get
-    @percentage = ((count.to_i / config.max_faltas.to_i) * 100).round(2)
+    @percentage = ((count.to_f / config.max_faltas.to_f) * 100).round(2)
   end
 
   def graph
